Add ecdsa-based key reconstructor
This is an alternative to unfactor.py, which works with files of unknown magic number,
and can recover both kinds of private keys used by TeslaCrypt.
(Bitcoin private keys can be used with TeslaDecoder or TeslaDecrypt.)
The sample file name and factors are specified on the command line in this version.

The downside is that this version runs a lot slower than the original - however, it's
guaranteed not to produce false positive keys, i.e. the key revealed by this script
is always the actual key, not just a candidate.